+++ R/storage.r
@@ -20,15 +20,8 @@ sbgj_storage_all <- function(output = NULL, ..., paper = c("letter", "a4"), inst
 	paper <- match.arg(paper)
 	output <- pnpmisc:::normalize_output(output)
 
-	bm <- data.frame(
-		title = c("Dice", "Glass Stones", "Pawns", "Polyhedral Dice"),
-		page = c(1L, 3L, 5L, 7L)
-	)
-	if (isTRUE(instructions)) {
-		bmi <- data.frame(title = "Instructions", page = 1L)
-		bm$page <- bm$page + 2L
-		bm <- rbind(bmi, bm)
-	}
+	title <- c("Dice", "Glass Stones", "Pawns", "Polyhedral Dice")
+	bm <- bm_from_title(title, instructions)
 
 	xmp <- xmp(
 		creator = "Trevor L. Davis",